requirements.txt
flask==2.3.3
flask-cors==4.0.0
agno==0.1.0
google-generativeai==0.3.2
anthropic==0.25.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
